fix: do not commit docs updates on sitemap only
Don't commit documentation updates if only the sitemap changed. This is
likely to happen when rebuilding the mkdocs site on a day that's
different from when the site was last built. In this case, the only
change that mkdocs applies is to refresh the sitemap last update date,
causing unnecessary diffs, and potential conflicts.